Take addons to fetch from JSON file
With this change the addons no longer have to be specified in the
source code, instead the program takes a JSON formatted file
containing the addon slugs as input.

This change also removes the previous generation of an overlay.
Instead the project just contains the tool and the actual daily
generation of a package set is handled in

    https://gitlab.com/rycee/nur-expressions